import clock widget directly
