+++ ui/site/src/site.ts
@@ -3,7 +3,6 @@ import Mousetrap from './mousetrap';
 import { spinnerHtml } from 'common/spinner';
 import { randomToken } from 'common/algo';
 import powertip from './powertip';
-import clockWidget from './clockWidget';
 import * as assets from './asset';
 import makeLog from './log';
 import pubsub from './pubsub';
@@ -24,7 +23,6 @@ const s = window.site;
 s.mousetrap = new Mousetrap(document);
 s.sri = randomToken();
 s.powertip = powertip;
-s.clockWidget = clockWidget;
 s.spinnerHtml = spinnerHtml;
 s.asset = assets;
 s.pubsub = pubsub;
